JWT token creted while after login and authentication and authorization done

+++ controllers/api/v1/post_api.js
@@ -22,7 +22,6 @@ module.exports.index = async function(req, res) {
 module.exports.destroy = async function(req ,res){
   
     try{
-
       //berore deleing a post we need to check whether particular post is exist or not from db
       const post = await Post.findById(req.params.id); // getting id of post through string param
 
@@ -30,15 +29,16 @@ module.exports.destroy = async function(req ,res){
       // postman via its appropritae URL , so i.e why in req user info will not present(req.user.id
       //  will give undefined)
       // so we commented below line and removed req.flash lines
-
-      //if(post.user == req.user.id){
+      // now we are applying authentication and authorization for deleting a post
+      if(post.user == req.user.id){
 
         // till here we are authencticating user vis broswer and stores user info in cookie file
-      // in next will authenticate user via postman using JWT because, API's do not have cookies
+      // in next will authenticate user via postman using JWT because, as API's do not have cookies
          await post.deleteOne(); // post.remove() is not working in newer version
   
          // and then delete all the comments associated with that post
          const deletedComments = await Comment.deleteMany({post: req.params.id});
+         console.log("Deleted Commnets: ",deletedComments);
          if(deletedComments) {
             //req.flash('success', "Post and associated comments deleted successfully");
           return res.json(200, {
@@ -47,6 +47,12 @@ module.exports.destroy = async function(req ,res){
          }
         
       }
+      else{
+        return res.json(401, {
+          message: "You cannot delete this post"
+        });
+      }
+    }
     catch(err){
             return res.json(500, {
                 message: "Internal Server Error"
